Changing names of projects and removing non necessary usings
