--- Inventory/Program.cs
@@ -1,2 +1,0 @@
-﻿
-
