+++ TheUncodedOne.Contract/Interfaces/IAction.cs
@@ -1,0 +1,7 @@
+﻿namespace TheUncodedOne.Contract.Interfaces;
+
+public interface IAction
+{
+    void Run(IBattle battle, ICharacter character);
+}
+
