+++ TheUncodedOne.Contract/Interfaces/IAttack.cs
@@ -1,0 +1,7 @@
+﻿namespace TheUncodedOne.Contract.Interfaces;
+public record AttackData(int Damage);
+public interface IAttack
+{
+    string Name { get; }
+    AttackData Create();
+}
