+++ TheUncodedOne.Contract/Interfaces/IBattle.cs
@@ -1,0 +1,6 @@
+﻿namespace TheUncodedOne.Contract.Interfaces;
+
+public interface IBattle
+{
+	public IParty GetEnemyPartyFor(ICharacter character);
+}
