+++ TheUncodedOne.Contract/Interfaces/ICharacter.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TheUncodedOneContract.Interfaces;
+﻿namespace TheUncodedOne.Contract.Interfaces;
 
 public interface ICharacter
 {
