+++ TheUncodedOne.Contract/Interfaces/IDisplay.cs
@@ -1,0 +1,7 @@
+﻿namespace TheUncodedOne.Contract.Interfaces;
+
+public interface IDisplay
+{
+	void DisplayClear();
+	void DisplayText(string text, ConsoleColor color = ConsoleColor.White);
+}
