+++ TheUncodedOne.Contract/Interfaces/IItem.cs
@@ -1,0 +1,8 @@
+﻿namespace TheUncodedOne.Contract.Interfaces;
+
+public record ItemData(int Points);
+public interface IItem
+{
+	string Name { get; }
+	ItemData Create();
+}
