+++ TheUncodedOne.Contract/Interfaces/IParty.cs
@@ -1,0 +1,8 @@
+﻿namespace TheUncodedOne.Contract.Interfaces;
+
+public interface IParty
+{
+	public IPlayer Player { get; }
+	public List<ICharacter> Characters { get; }
+	public IDisplay Display { get; init; }
+}
