+++ TheUncodedOne.Contract/Interfaces/IPlayer.cs
@@ -1,0 +1,6 @@
+﻿namespace TheUncodedOne.Contract.Interfaces;
+
+public interface IPlayer
+{
+    IAction ChooseAction(IBattle battle, ICharacter character, int typeAction);
+}
