+++ TheUncodedOne.Game/Actions/AttackAction.cs
@@ -1,9 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-using TheUncodedOne.Contract.Interfaces;
+﻿using TheUncodedOne.Contract.Interfaces;
 using TheUncodedOne.Game.Attacks;
 using TheUncodedOne.Game.Characters;
 using TheUncodedOne.Game.Displays;
