+++ TheUncodedOne.Game/Characters/TheUncodedOneCharacter.cs
@@ -1,5 +1,6 @@
-﻿using TheUncodedOne.Contract.Interfaces;
-using TheUncodedOneGame.Attacks;
+﻿using System;
+using TheUncodedOne.Contract.Interfaces;
+using TheUncodedOne.Game.Attacks;
 
 namespace TheUncodedOne.Game.Characters;
