+++ TheUncodedOne.Game/Game.cs
@@ -8,14 +8,14 @@ using TheUncodedOne.Game.Characters;
 using TheUncodedOne.Game.Displays;
 using TheUncodedOne.Game.Players;
 
-namespace TheUncodedOneGame;
+namespace TheUncodedOne.Game;
 
 public class Game
 {
 	private List<Party> Monsters { get; set; }
 	private Party Heroes { get; set; }
 	//private int GameMode { get; set; }
-	private IDisplay _display;
+	private readonly IDisplay _display;
 	private Mode Mode;
 	private void InitializePlayerName(Party heroes)
 	{
