+++ TheUncodedOne.Game/TheUncodedOne.Game.csproj
@@ -8,7 +8,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <ProjectReference Include="..\TheUncodedOneContract\TheUncodedOneContract.csproj" />
+    <ProjectReference Include="..\TheUncodedOne.Contract\TheUncodedOne.Contract.csproj" />
   </ItemGroup>
 
 </Project>
