+++ TheUncodedOne.Inventory/Items/HealthPotionItem.cs
@@ -1,0 +1,10 @@
+﻿using TheUncodedOne.Contract.Interfaces;
+
+namespace ITheUncodedOne.Inventory.Actions;
+
+public class HealthPotionItem : IItem
+{
+	private readonly int _healthPoints = 10;
+	public string Name => "Health potion";
+	public ItemData Create() => new(_healthPoints);
+}
