+++ TheUncodedOne.Inventory/TheUncodedOne.Inventory.csproj
@@ -1,10 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>net8.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
     <Nullable>enable</Nullable>
   </PropertyGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\TheUncodedOne.Contract\TheUncodedOne.Contract.csproj" />
+  </ItemGroup>
 
 </Project>
