--- TheUncodedOneContract/Interfaces/IAttack.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TheUncodedOneContract.Interfaces;
-public record AttackData(int Damage);
-public interface IAttack
-{
-    string Name { get; }
-    AttackData Create();
-}
