--- TheUncodedOneContract/Interfaces/IBattle.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TheUncodedOneContract.Interfaces;
-
-public interface IBattle
-{
-	public IParty GetEnemyPartyFor(ICharacter character);
-}
