--- TheUncodedOneContract/Interfaces/IDisplay.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace TheUncodedOneContract.Interfaces;
-
-public interface IDisplay
-{
-	void DisplayClear();
-	void DisplayText(string text, ConsoleColor color = ConsoleColor.White);
-}
