--- TheUncodedOneContract/Interfaces/IPlayer.cs
@@ -1,13 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-
-namespace TheUncodedOneContract.Interfaces;
-
-public interface IPlayer
-{
-    IAction ChooseAction(IBattle battle, ICharacter character, int typeAction);
-}
